[ios][expo-network] refactor
